Added workaround for providers specified in state
I think this is due to version changes over time.

I found many related issues with upgrading to new versions of Terraform.

There is also a command `terraform state replace-provider`, but the
source of the problem seems to be that `init` only takes care of the
default workspace. So, running `init`, switching workspaces and then
running `init` again also seems to fix this.